requirements.txt
Flask==3.0.3
python-dotenv
pymongo[srv]==4.6.3
gunicorn
